Tokenize comments
Comment style is an important syntactic feature of programming
languages. This commit changes the tokenizer to output a special token
for each comment.

A comment such as "/* Comment */" will generate the token "COMMENT/*".

This already improves detection of some sames. For example, if
Mathematica/PacletInfo.m is excluded from training, it will be
classified as Mercury instead of Mathematica (on
test_classify_ambiguous_languages), but after this commit, it will be
classified correctly, because the "COMMENT(*" tokens provide enough
information to the classifier to make the correct decision.

Tests are adjusted to:

- Ignore classifier tests for extensions where we have a catch-all rule.
This includes .sql, where the classifier output is mostly useless.

- Allow `samples/C++/rpc.h` to be classified as C. This file is fully
valid C without any single C++-specific syntax or convention.